src/quantum_consciousness/core/quantum_topology.py
"""
Quantum topology implementation for TQFT and quantum homology calculations.
Implements cobordism maps, quantum homology, and topological invariants.
"""

import numpy as np
from typing import List, Tuple, Dict, Optional
from scipy.linalg import expm
import networkx as nx

class QuantumTopology:
    """Implements quantum topological operations and TQFT calculations."""

    def __init__(self, dim: int):
        """
        Initialize quantum topology calculator.

        Args:
            dim: Dimension of the quantum system
        """
        self.dim = dim
        self.cobordism_maps = {}
        self.homology_groups = {}
        self.quantum_invariants = {}

    def create_cobordism(self, initial_manifold: np.ndarray,
                        final_manifold: np.ndarray) -> np.ndarray:
        """
        Create cobordism between quantum manifolds
        Z(M): TQFTₙ → Vect

        Args:
            initial_manifold: Initial quantum state manifold
            final_manifold: Final quantum state manifold

        Returns:
            numpy.ndarray: Cobordism map
        """
        # Calculate dimension of cobordism
        dim_cobordism = len(initial_manifold) + len(final_manifold)

        # Create cobordism matrix
        cobordism = np.zeros((dim_cobordism, dim_cobordism), dtype=complex)

        # Implement gluing map
        for i in range(len(initial_manifold)):
            for j in range(len(final_manifold)):
                cobordism[i,j] = np.vdot(initial_manifold[i], final_manifold[j])

        return cobordism

    def compute_quantum_homology(self, quantum_complex: List[np.ndarray]) -> Dict[int, np.ndarray]:
        """
        Compute quantum homology groups
        H_n(X) = ker(∂_n)/im(∂_{n+1})

        Args:
            quantum_complex: List of boundary maps

        Returns:
            Dict[int, numpy.ndarray]: Homology groups
        """
        homology_groups = {}

        for n in range(len(quantum_complex) - 1):
            # Calculate kernel of boundary map n
            kernel_n = np.linalg.null_space(quantum_complex[n])

            # Calculate image of boundary map n+1
            image_n_plus_1 = quantum_complex[n+1]

            # Compute quotient space
            homology_n = np.linalg.qr(np.hstack((kernel_n, image_n_plus_1)))[0]
            homology_groups[n] = homology_n

        return homology_groups

    def calculate_jones_polynomial(self, braiding_matrix: np.ndarray) -> np.polynomial.polynomial.Polynomial:
        """
        Calculate Jones polynomial invariant
        V_L(t) = (-A³)^{w(L)} ⟨L⟩

        Args:
            braiding_matrix: Matrix representing braiding operations

        Returns:
            numpy.polynomial.polynomial.Polynomial: Jones polynomial
        """
        # Calculate writhe number
        writhe = np.trace(braiding_matrix)

        # Initialize Kauffman bracket
        A = np.exp(2j * np.pi / 8)  # A = e^{2πi/8}
        kauffman = 0

        # Calculate Kauffman bracket polynomial
        for i in range(self.dim):
            kauffman += np.linalg.matrix_power(A, 3*writhe) * np.trace(
                np.linalg.matrix_power(braiding_matrix, i))

        return np.polynomial.polynomial.Polynomial(kauffman)

    def compute_tqft_partition(self, manifold: np.ndarray) -> complex:
        """
        Compute TQFT partition function
        Z(M) = ∫ DA e^{iS[A]}

        Args:
            manifold: Quantum manifold

        Returns:
            complex: Partition function value
        """
        # Calculate action
        S = np.trace(manifold @ manifold.conj().T)

        # Compute partition function
        Z = np.exp(1j * S)

        return Z

    def calculate_quantum_invariants(self, state: np.ndarray) -> Dict[str, float]:
        """
        Calculate quantum topological invariants

        Args:
            state: Quantum state

        Returns:
            Dict[str, float]: Dictionary of invariants
        """
        invariants = {}

        # Calculate first Chern number
        density_matrix = np.outer(state, state.conj())
        berry_curvature = np.imag(np.log(np.trace(density_matrix)))
        invariants['chern_number'] = berry_curvature / (2 * np.pi)

        # Calculate winding number
        phase = np.angle(state)
        winding = np.sum(np.diff(phase)) / (2 * np.pi)
        invariants['winding_number'] = winding

        return invariants

    def construct_link_invariant(self, braid_word: List[int]) -> float:
        """
        Construct link invariant from braid word

        Args:
            braid_word: List of integers representing braid generators

        Returns:
            float: Link invariant value
        """
        # Initialize braid representation
        representation = np.eye(self.dim, dtype=complex)

        # Standard generators of braid group
        sigma_1 = np.array([[0, 1], [1, 0]], dtype=complex)
        sigma_2 = np.array([[1, 0, 0], [0, 0, 1], [0, 1, 0]], dtype=complex)

        # Construct braid representation
        for generator in braid_word:
            if generator > 0:
                representation = representation @ sigma_1
            else:
                representation = representation @ np.linalg.inv(sigma_2)

        # Calculate trace of representation
        return abs(np.trace(representation))

    def compute_quantum_cohomology(self, quantum_complex: List[np.ndarray]) -> Dict[int, np.ndarray]:
        """
        Compute quantum cohomology rings
        H^*(X) = ⊕ H^n(X)

        Args:
            quantum_complex: List of coboundary maps

        Returns:
            Dict[int, numpy.ndarray]: Cohomology groups
        """
        cohomology_groups = {}

        for n in range(len(quantum_complex) - 1):
            # Calculate kernel of coboundary map
            kernel = np.linalg.null_space(quantum_complex[n].T)

            # Calculate image of previous coboundary map
            if n > 0:
                image = quantum_complex[n-1].T
            else:
                image = np.zeros_like(kernel)

            # Compute cohomology group
            cohomology_n = np.linalg.qr(np.hstack((kernel, image)))[0]
            cohomology_groups[n] = cohomology_n

        return cohomology_groups

    def calculate_spectral_sequence(self, filtration: List[np.ndarray]) -> Dict[Tuple[int, int], np.ndarray]:
        """
        Calculate spectral sequence for filtered complex
        E_r^{p,q} ⇒ H^{p+q}(X)

        Args:
            filtration: List of filtered subcomplexes

        Returns:
            Dict[Tuple[int, int], numpy.ndarray]: Spectral sequence pages
        """
        spectral_sequence = {}

        for r in range(len(filtration)):
            for p in range(len(filtration[r])):
                for q in range(len(filtration[r])):
                    if p + q < len(filtration[r]):
                        # Calculate differential d_r
                        d_r = filtration[r][p+r] @ filtration[r][q]

                        # Calculate cohomology of d_r
                        kernel = np.linalg.null_space(d_r)
                        image = d_r @ filtration[r][p+q]

                        # Store result in spectral sequence
                        spectral_sequence[(p,q)] = np.linalg.qr(
                            np.hstack((kernel, image)))[0]

        return spectral_sequence
